feat: forester: grpc processing for v2 trees (#2024)
* feat(forester): gRPC-based event-driven processing for V2 trees

* chore: add protobuf-compiler to setup-and-build

* chore: update PHOTON_COMMIT version in versions.sh

* feat: add grpc_port config to cli

* cleanup

* cleanup

* cleanup

* wait for indexer in rpc-interop.test.ts

* wait for indexer in rpc-interop.test.ts

* bump photon version

* cleanup

* cleanup

+++ sdk-libs/client/src/indexer/photon_indexer.rs
@@ -1655,7 +1655,7 @@ impl Indexer for PhotonIndexer {
                                 value: QueueElementsResult {
                                     elements: proofs,
                                     first_value_queue_index: Some(
-                                        api_result.first_value_queue_index as u64,
+                                        api_result.first_value_queue_index,
                                     ),
                                 },
                             })
